backend/requirements.txt
# FastAPI y servidor
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.1

# LangChain y LangGraph - Versiones compatibles con Python 3.13
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-ollama>=0.2.0

# Vector Store
chromadb>=0.5.0

# Document Processing
pypdf>=5.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Embeddings
sentence-transformers>=3.0.0

# Utilities (versiones nuevas con soporte Python 3.13)
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.27.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0